split out command/args/mods in ExCmd nodes
I don't want to introduce a million individual Node variants in the
parser, so I'm going with a different strategy. Like the original VimL
parser from which this came, all commands will be parsed as ExCmd nodes,
and anything using the parser lib will have to parse those further if so
desired. However, I am splitting out the mods, bang, and command from
the args, so that much is done in advance.

So for example:

    silent highlight! default Normal guifg=#123456

will parse as an ExCmd with these properties:

- mods: ["silent"]
- command: "highlight"
- bang: true
- args: "default Normal guifg=#123456"

At least this way if the user wishes to parse the command further, the
args have already been separated, and they can match based on the
command name.

There were also a couple minor bugs I found and fixed that were
unrelated to this change.